Plugin manager refactoring for proper releases
