+++ src/aye/plugin_base.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Optional
 
 class Plugin(ABC):
     name: str
-    version: str = "0.1.0"
+    version: str = "1.0.0"
     premium: str = "free"  # one of: free, pro, team, enterprise
 
     @abstractmethod
