+++ src/aye/plugins/.gitignore
@@ -1,0 +1,2 @@
+plugin_base.py
+
